clean up shim
